+++ app/src/main/java/com/LaursenJessen/bikeshare/firestore/Bike.kt
@@ -1,11 +1,13 @@
 package com.LaursenJessen.bikeshare.firestore
 
+import android.os.Parcelable
+
 data class Bike(
     val id:String,
     val address:String,
     val description:String,
     val name:String,
-    val distance:Float,
+    val distance:Int,
     val rentedOut:Boolean,
     val userId:String,
     val imageUrl:String
